Organze tests and parallelize them by route
