+++ gradle.properties
@@ -1,1 +1,2 @@
 version = 0.3.1-SNAPSHOT
+org.gradle.parallel
